LPS-115572 | LPS-112434 Move WorkflowInfoBar to shared components
