+++ modules/dxp/apps/app-builder-workflow/app-builder-workflow-web/src/main/resources/META-INF/resources/css/WorkflowInfoBar.scss
@@ -1,4 +1,4 @@
-.workflow-info {
+.workflow-info-bar {
 	display: flex;
 	justify-content: center;
